feat: refactor filteringLogic
