+++ api/src/test/java/demo/BaseIntegrationTest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import demo.controller.dto.SaveUserDto;
 import demo.controller.dto.SectorDto;
 import demo.controller.dto.UserFilterDto;
-import demo.service.validation.filter_validator.FilterService;
+import demo.service.FilterService;
 import demo.service.SectorService;
 import demo.service.UserService;
 import jakarta.persistence.EntityManager;
