+++ api/src/test/java/demo/service/filter/FilteringLogicServiceTest.java
@@ -1,0 +1,14 @@
+package demo.service.filter;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class FilteringLogicServiceTest {
+
+  @Test
+  void findAllByFilter() {
+
+
+  }
+}
